Learned about Data.Unique and fixed the ID system
I spent quite some time trying to figure out how to solve this
problem. I needed a monad and an "ID" type that cannot escape
monad. This was an important restriction: if the ID could escape the
monad, it would be impossible to guarantee that the ID is unique!  For
instance, someone could run a world computation for a while, then take
an ID from that world and attempt to use it in a new world.

I first considered ST, since the restriction "an STRef cannot escape ST"
is the entire reason ST works. However:
- I didn't want to write a little 's' everywhere, and
- STRefs do not implement Ord.

The former problem could have been solved by considering IO/IORef, but
that still does not implement Ord, meaning I would not be able to use
these as keys in a Map. I considered how to solve this by defining
an ID monad like

newtype ID a = ID (StateT Integer IO a)

but it seemed that to implement it, I needed to pick between letting
my "IDRef"s escape the ID monad or being able to convert ID to IO.

I'm so glad I found Data.Unique. The fact that it is in the base
package really validates my usage of it!